Fix log printing of before/after cleanup
Seems like MinGW doesn't have the %z (size_t) printf format
So add some %f casting variant, so that it works everywhere.

+++ src/cleanup.c
@@ -173,13 +173,13 @@ clean_up_dir(struct conf *conf, const char *dir, float limit_multiple)
 	traverse(dir, traverse_fn);
 
 	// Clean the cache.
-	cc_log("Before cleanup: %lu KiB, %zu files",
+	cc_log("Before cleanup: %lu KiB, %.0f files",
 	       (unsigned long)cache_size / 1024,
-	       files_in_cache);
+	       (float)files_in_cache);
 	bool cleaned = sort_and_clean();
-	cc_log("After cleanup: %lu KiB, %zu files",
+	cc_log("After cleanup: %lu KiB, %.0f files",
 	       (unsigned long)cache_size / 1024,
-	       files_in_cache);
+	       (float)files_in_cache);
 
 	if (cleaned) {
 		cc_log("Cleaned up cache directory %s", dir);
